requirements.txt
editdistance==0.8.1
fairseq==0.12.2
ffmpeg_python==0.2.0
hydra-core==1.3.2
ibug_face_alignment.egg==info
ibug_face_detection.egg==info
nltk==3.8.1
numpy==1.24.3
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
python_Levenshtein==0.25.0
python_Levenshtein==0.25.0
pytorch_lightning==1.5.10
sentencepiece==0.1.99
six==1.16.0
skimage==0.0
torch==2.1.1
torch_complex==0.4.3
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
warpctc_pytorch==0.2.1
whisper==1.1.10
